ref: Extracts BridgeState to a separate file.
